--- keep-ui/app/(keep)/workflows/create-workflow-modal/index.ts
@@ -1,2 +1,0 @@
-export { CreateWorkflowModal } from "./create-workflow-modal";
-export { WorkflowTemplates } from "./workflow-templates";
