+++ keep-ui/app/(keep)/workflows/workflow-templates/index.ts
@@ -1,0 +1,1 @@
+export { WorkflowTemplates } from "./workflow-templates";
