+++ keep-ui/app/(keep)/workflows/workflow-templates/workflow-template-card.tsx
